A-01: каталог товаров — GET /products, GET /products/:id, поиск и фил…

+++ internal/model/order_item.go
@@ -1,9 +1,9 @@
 package model
 
 type OrderItem struct {
-	ID        int64   `json:"id"`
-	OrderID   int64   `json:"order_id"`
-	ProductID int64   `json:"product_id"`
-	Quantity  int64   `json:"quantity"`
-	Price     float64 `json:"price"`
+	ID        int64 `json:"id"`
+	OrderID   int64 `json:"order_id"`
+	ProductID int64 `json:"product_id"`
+	Quantity  int64 `json:"quantity"`
+	Price     int64 `json:"price"`
 }
